requirements.txt
kafka-python==2.0.2
prometheus_client==0.20.0
psycopg2-binary==2.9.9
python-json-logger==2.0.7
fastapi==0.111.0
uvicorn==0.30.1
